# Make static file reading (content/metadata) operations sync