@bjornharrtell 2022-10-17
Very interesting! It suprises me quite a bit that rust async framework has such a large overhead.