@joseluisq 2022-10-18
> Very interesting! It suprises me quite a bit that rust async framework has such a large overhead.

As far as I can tell, it's not only because of Tokio runtime but the `async` usage in particular for our io file reads since those operations really happen `sync` under the hood (blocking system calls), and using Tokio via `tokio::fs` adds extra overhead. That's why our old `file_stream` implementation was definitely involved (as removed in the PR).

But here is a remarkable explanation from Tokio website talking about Linux `io_uring` support.

> All `tokio-uring` operations are truly `async`, **unlike APIs provided by `tokio::fs`, which run on a thread pool. Using synchronous filesystem operations from a thread pool adds significant overhead.** With `io-uring`, we can perform both network and file system operations asynchronously from the same thread. But, io-uring is a lot more.
> https://tokio.rs/blog/2021-07-tokio-uring

By the way, I have some ideas in mind to add experimental support for _really asynchronous (non-blocking) file read operations_ using `io_uring` in Linux via the [tokio-uring](https://github.com/tokio-rs/tokio-uring) crate soon.